requirements.txt
streamlit==1.32.0
pandas==2.2.1
python-dotenv==1.0.0 